feature: create radioBlock component with styles
