+++ src/features/profile/ui/settings/accountManagement/accountManagement.module.scss
@@ -8,10 +8,10 @@
   height: 100%;
   margin-top: 20px;
 
-  & .radio {
-    width: 15px;
-    height: 15px;
-  }
+  // & .radio {
+  //   width: 15px;
+  //   height: 15px;
+  // }
 }
 
 // .renewal {
@@ -66,10 +66,10 @@
   }
 }
 
-.block {
-  display: flex;
-  gap: 8px;
-}
+// .block {
+//   display: flex;
+//   gap: 8px;
+// }
 
 .title {
   margin-bottom: 5px;
