+++ src/features/profile/ui/settings/accountManagement/radioBlock/index.ts
@@ -1,0 +1,1 @@
+export * from './radioBlock'
