+++ src/features/profile/ui/settings/accountManagement/radioBlock/radioBlock.module.scss
@@ -1,0 +1,9 @@
+.radio {
+  width: 15px;
+  height: 15px;
+}
+
+.block {
+  display: flex;
+  gap: 8px;
+}
